# Implement schedule simulator in the fuzzer, further performance improvements and bug fixes.