--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,9 +14,8 @@ license = "MIT/Apache-2.0"
 debug = 2
 
 [dependencies]
-derivative = "1.0"
+fnv = "1.0"
 gfx-hal = { version = "0.1.0", git = "https://github.com/gfx-rs/gfx", rev = "1e959ace6" }
-log = "0.4"
 
 [dev-dependencies]
 clap = "2.31"