--- a/src/resource/access.rs
+++ b/src/resource/access.rs
@@ -1,9 +1,11 @@
 use std::fmt::Debug;
-use std::ops::{BitOr, BitOrAssign};
+use std::ops::{BitAnd, BitAndAssign, BitOr, BitOrAssign};
 use hal::pso::PipelineStage;
 
 /// Access type combination
-pub trait Access: Debug + Copy + BitOr<Output = Self> + BitOrAssign + Eq {
+pub trait Access: Debug + Copy + Eq +
+                  BitAnd<Output = Self> + BitAndAssign +
+                  BitOr<Output = Self> + BitOrAssign {
     /// Create empty combinations of access types.
     fn none() -> Self;