This implements a schedule simulator that simulates random execution orders on a schedule in an attempt to find issues, and fixes several bugs caught by it. 

In addition, it make changes taking advantage of a few additional optimization opportunities, especially when only a single queue is in use.

Before:
```
tiny   + single-queue: 101483 iters ran in average 0.027 ± 0.008 ms (0.16% ± 0.05% of 16.6 ms)
tiny   + multi-queue :  63530 iters ran in average 0.050 ± 0.007 ms (0.30% ± 0.04% of 16.6 ms)
small  + single-queue:  15568 iters ran in average 0.195 ± 0.019 ms (1.17% ± 0.11% of 16.6 ms)
small  + multi-queue :   8918 iters ran in average 0.384 ± 0.050 ms (2.32% ± 0.30% of 16.6 ms)
medium + single-queue:   4605 iters ran in average 0.684 ± 0.080 ms (4.12% ± 0.48% of 16.6 ms)
medium + multi-queue :   2795 iters ran in average 1.248 ± 0.073 ms (7.52% ± 0.44% of 16.6 ms)
large  + single-queue:    293 iters ran in average 12.218 ± 0.440 ms (73.60% ± 2.65% of 16.6 ms)
large  + multi-queue :    182 iters ran in average 20.427 ± 0.705 ms (123.05% ± 4.25% of 16.6 ms)
huge   + single-queue:    162 iters ran in average 22.301 ± 0.904 ms (134.34% ± 5.44% of 16.6 ms)
huge   + multi-queue :    105 iters ran in average 35.773 ± 0.985 ms (215.50% ± 5.93% of 16.6 ms)
```

After:
```
tiny   + single-queue: 124558 iters ran in average 0.019 ± 0.003 ms (0.12% ± 0.02% of 16.6 ms)
tiny   + multi-queue :  70690 iters ran in average 0.043 ± 0.006 ms (0.26% ± 0.04% of 16.6 ms)
small  + single-queue:  19696 iters ran in average 0.138 ± 0.013 ms (0.83% ± 0.08% of 16.6 ms)
small  + multi-queue :   9779 iters ran in average 0.343 ± 0.032 ms (2.06% ± 0.19% of 16.6 ms)
medium + single-queue:   6166 iters ran in average 0.456 ± 0.047 ms (2.75% ± 0.28% of 16.6 ms)
medium + multi-queue :   3148 iters ran in average 1.078 ± 0.066 ms (6.49% ± 0.40% of 16.6 ms)
large  + single-queue:    461 iters ran in average 6.989 ± 0.592 ms (42.10% ± 3.57% of 16.6 ms)
large  + multi-queue :    226 iters ran in average 15.993 ± 0.660 ms (96.34% ± 3.98% of 16.6 ms)
huge   + single-queue:    293 iters ran in average 11.099 ± 0.541 ms (66.86% ± 3.26% of 16.6 ms)
huge   + multi-queue :    135 iters ran in average 26.934 ± 0.881 ms (162.25% ± 5.31% of 16.6 ms)
```