@Lymia 2018-06-05
I can manually check that the result makes sense, I guess, but this will shuffle change the ordering of passes chosen with a single queue, I'd imagine.