@Lymia 2018-06-12
Sorry for the delay. I fixed the issue with `queue_state`.